LayoutTests/media/session/track-media-events-in-content-sessions-expected.txt

'Content' sessions that are active should respond to Previous/Next Track media control events.


EVENT(canplaythrough)
EXPECTED (internals.mediaSessionCurrentState(session) == 'idle') OK
EXPECTED (video.paused == 'true') OK
Begin playing media.
EVENT(playing)
Media began playing.
EXPECTED (video.paused == 'false') OK
EXPECTED (internals.mediaSessionCurrentState(session) == 'active') OK
Sending Next Track media event.
RUN(internals.sendMediaControlEvent("next-track"))
EVENT(nexttrack)
Sending Previous Track media event.
RUN(internals.sendMediaControlEvent("previous-track"))
EVENT(previoustrack)
END OF TEST
